% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getpop}
\alias{getpop}
\title{Get population for one or more localities}
\usage{
getpop(locality = NULL, fips = NULL)
}
\arguments{
\item{locality}{Name of localities to get population for.  Ignored if \code{fips}
is specified}

\item{fips}{FIPS codes for localities to get populaiton for.  If specified this
overrides the \code{locality} parameter}
}
\value{
Vector of populations
}
\description{
Get population for one or more localities
}
