% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vadailycases}
\alias{vadailycases}
\title{Daily cumulative cases by locality}
\format{
Data frame with 8 columns
\describe{
\item{date}{Date of the report.}
\item{fips}{FIPS code for the locality.}
\item{locality}{Name of the locality (city or county).}
\item{county}{Alternate format for the name of the locality.}
\item{HealthDistrict}{VDH health district the locality belongs to.}
\item{cases}{Cumulative COVID-19 cases in the locality.}
\item{hosp}{Cumulative hospitalizations for COVID-19 in the locality.}
\item{deaths}{Cumulative deaths from COVID-19 in the locality.}
}
}
\source{
Virginia Department of Health
\url{https://www.vdh.virginia.gov/coronavirus/}
}
\usage{
vadailycases
}
\description{
Raw data from VDH.  Note that it starts on 17 March, which is a few weeks later
than the testing data.
}
\keyword{datasets}
