% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vadailytests}
\alias{vadailytests}
\title{Daily VDH COVID-19 testing data by health district}
\format{
Data frame with 4 columns
\describe{
\item{date}{Date of the lab reports.}
\item{HealthDistrict}{VDH health district.}
\item{ntest}{Number of PCR test encounters reported.}
\item{npos}{Number of positive tests reported.}
}
}
\source{
Virginia Department of Health
\url{https://www.vdh.virginia.gov/coronavirus/}
}
\usage{
vadailytests
}
\description{
This is the raw VDH testing and positive case fraction reports by health
district.
}
\keyword{datasets}
