% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vadailyvax}
\alias{vadailyvax}
\title{Daily vaccinations by locality}
\format{
Data frame with 8 columns
\describe{
\item{date}{Date of the observation}
\item{fips}{FIPS code for the locality}
\item{locality}{Name of the locality}
\item{population}{Population of the locality}
\item{district}{Local health district of the locality}
\item{ndose}{Number of doses administered. Zero dose days are recorded explicitly.}
\item{vaxtotal}{Cumulative number of doses administered}
\item{vaxfrac}{Cumulative fraction of the population vaccinated.}
}
}
\usage{
vadailyvax
}
\description{
Number of \emph{final dose} vaccinations (total over all manufacturers)
administered, by locality and date.
}
\keyword{datasets}
