% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{valocalities}
\alias{valocalities}
\title{Virginia locality characteristics}
\format{
Data frame with 8 columns
\describe{
\item{fips}{FIPS code for the locality}
\item{county}{Name of the locality}
\item{district}{VDH health district the locality belongs to}
\item{areaSqMile}{Area of the locality in square miles}
\item{population}{Population of the locality}
\item{locality}{Alternate form of the locality name.  This is sometimes useful
for joining to tables from VA government sources.}
\item{popdens}{Average population density in the locality.}
\item{stdpopdens}{Standardized average population density in the locality.}
}
}
\usage{
valocalities
}
\description{
Provide static data such as population and population density for Virginia
localities.  While not technically related to COVID-19, these statistics are
sometimes useful in modeling.
}
\keyword{datasets}
