% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vaweeklycases}
\alias{vaweeklycases}
\title{Weekly new cases by locality}
\format{
Data frame with 8 columns
\describe{
\item{date}{Date of the last day of the week.}
\item{fips}{FIPS code for the locality.}
\item{locality}{Name of the locality (city or county).}
\item{HealthDistrict}{VDH health district the locality belongs to.}
\item{week}{Week number, counting from the beginning of the year.}
\item{cases}{New cases in the locality for the week.}
\item{hosp}{New hospitalizations in the locality for the week.}
\item{deaths}{New deaths in the locality for the week.}
}
}
\source{
Virginia Department of Health
\url{https://www.vdh.virginia.gov/coronavirus/}
}
\usage{
vaweeklycases
}
\description{
This is aggregated from the daily data.  We give new cases for each week
instead of cumulative cases.  As with \code{\link{vaweeklytests}}, the date
is given for the last day of the week.
}
\keyword{datasets}
